Fix real libmpv smoke WAV fixture

+++ desktopApp/src/test/kotlin/org/feeluown/mobile/desktop/DesktopMpvRealIntegrationTest.kt
@@ -85,7 +85,7 @@ class DesktopMpvRealIntegrationTest {
                 put("WAVE".encodeToByteArray())
                 put("fmt ".encodeToByteArray())
                 putInt(16)
-                putShort(1)
+                putShort(1.toShort())
                 putShort(channels.toShort())
                 putInt(sampleRate)
                 putInt(byteRate)
